Move DATE_FMT constant out of fab package
