+++ src/commcare_cloud/const.py
@@ -1,0 +1,1 @@
+DATE_FMT = '%Y-%m-%d_%H.%M'
